split java/build-tool generator
